fix: resolve 7 pre-existing test failures (Chantier 1)
All 7 failures on the default mvn test run are now green. Three were
real production bugs, four were stale tests against evolved code.

Production bugs:
 - StatsLogger.appendStats: piecesPerSec was always 0.0. Integer
   division (int depth × long MILLIS / long computeMs) truncated to 0
   whenever numerator < denominator (e.g. 48 × 1000 / 300000).
   Promote to double up front.
 - DomainManager.computeDomain: accepted null board silently when the
   pieces map was empty (the loop skipped every iteration and never
   dereferenced). Added Objects.requireNonNull at the boundary.
 - PuzzleFactory.loadEternityIIFromFile: rejected the on-disk Eternity
   II file because the parser didn't skip "# ..." comment headers and
   accepted only 4-token lines (the file ships with 5 tokens:
   pieceId N E S W, plus comment metadata). Support both shapes and
   skip comments.

Stale tests:
 - FormattingUtilsTest: captured System.out but FormattingUtils writes
   through SolverLogger. Switch to a Logback ListAppender for logger-
   based methods; keep System.setOut only for printProgressBar (uses
   \r + flush to repaint the terminal line).
 - BoardDisplayServiceTest.testEmptyCellsDisplayWithRotations: the
   setup placed two corner pieces whose edges could never fit the
   remaining cells, so countCandidatesWithRotations returned zero and
   the renderer emitted "∅" instead of "(p/r)". Rebuilt the setup with
   pieces that actually fit.
 - ValidPieceCounterTest.testCornerPieceRotations: the mock validator
   didn't enforce border rules (edge=0 on borders, edge≠0 interior),
   so a corner piece counted 4 rotations instead of 1. Tightened the
   mock to mirror the real PlacementValidator. Adjusted
   testMultipleValidPieces to target an interior cell accordingly.

Full suite: 1363 tests, 0 failures, 0 errors, 3 skipped (1m21s).

+++ src/main/java/util/StatsLogger.java
@@ -145,9 +145,12 @@ public class StatsLogger {
                 data.put("fitChecks", stats.fitChecks);
             }
 
-            // Calculate pieces per second
+            // Calculate pieces per second. Cast to double up-front: both operands
+            // were integer (depth is int, MILLIS_PER_SECOND and totalComputeTimeMs
+            // are long), so the division silently truncated to 0 whenever the
+            // numerator was smaller than the denominator (e.g. 48 * 1000 / 300000).
             if (totalComputeTimeMs > 0 && depth > 0) {
-                double piecesPerSec = (depth * TimeConstants.MILLIS_PER_SECOND) / totalComputeTimeMs;
+                double piecesPerSec = ((double) depth * TimeConstants.MILLIS_PER_SECOND) / totalComputeTimeMs;
                 data.put("piecesPerSec", Math.round(piecesPerSec * 100.0) / 100.0);
             } else {
                 data.put("piecesPerSec", 0.0);
